requirements.txt
streamlit==1.36.0
pandas==2.2.2
matplotlib==3.8.4
python-dotenv==1.0.1

# Azure Blob Storage
azure-storage-blob==12.19.1

# LangChain
langchain==0.2.11
langchain-community==0.2.6
langchain-core==0.2.24

# Chroma Vector Store
chromadb==0.5.0

# Sentence Transformers for embeddings
sentence-transformers==2.7.0
torch==2.2.2
transformers==4.41.2

# OpenAI LLM
openai==1.35.10
